Controller Refactoring st2
Did refactoring for controlles that were in the wrong classes.
