+++ Back-end (ISS)/Server-Back-ISS/src/main/java/ZgazeniSendvic/Server_Back_ISS/model/VehiclePosition.java
@@ -1,4 +1,4 @@
-package ZgazeniSendvic.Server_Back_ISS.module;
+package ZgazeniSendvic.Server_Back_ISS.model;
 
 public class VehiclePosition{
     private Long id;
